IGNITE-21886 Refactor CompressionProcessorImpl, move code partially to ignite-core module
