+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/wal/FileWriteAheadLogManager.java
@@ -161,8 +161,7 @@ import static org.apache.ignite.internal.processors.cache.persistence.wal.serial
 import static org.apache.ignite.internal.processors.cache.persistence.wal.serializer.RecordV1Serializer.HEADER_RECORD_SIZE;
 import static org.apache.ignite.internal.processors.cache.persistence.wal.serializer.RecordV1Serializer.readPosition;
 import static org.apache.ignite.internal.processors.cache.persistence.wal.serializer.RecordV1Serializer.readSegmentHeader;
-import static org.apache.ignite.internal.processors.compress.CompressionProcessor.checkCompressionLevelBounds;
-import static org.apache.ignite.internal.processors.compress.CompressionProcessor.getDefaultCompressionLevel;
+import static org.apache.ignite.internal.processors.compress.CompressionProcessor.getCompressionLevel;
 import static org.apache.ignite.internal.processors.configuration.distributed.DistributedBooleanProperty.detachedBooleanProperty;
 import static org.apache.ignite.internal.util.io.GridFileUtils.ensureHardLinkAvailable;
 
@@ -605,11 +604,10 @@ public class FileWriteAheadLogManager extends GridCacheSharedManagerAdapter impl
                         serializerVer);
                 }
 
-                cctx.kernalContext().compress().checkPageCompressionSupported();
+                if (pageCompression != DiskPageCompression.SKIP_GARBAGE)
+                    cctx.kernalContext().compress().checkPageCompressionSupported();
 
-                pageCompressionLevel = dsCfg.getWalPageCompressionLevel() != null ?
-                    checkCompressionLevelBounds(dsCfg.getWalPageCompressionLevel(), pageCompression) :
-                    getDefaultCompressionLevel(pageCompression);
+                pageCompressionLevel = getCompressionLevel(dsCfg.getWalPageCompressionLevel(), pageCompression);
             }
         }
     }
