+++ modules/core/src/test/java/org/apache/ignite/testsuites/IgnitePdsTestSuite5.java
@@ -50,6 +50,7 @@ import org.apache.ignite.internal.processors.cache.persistence.wal.SysPropWalDel
 import org.apache.ignite.internal.processors.cache.persistence.wal.WalArchiveConsistencyTest;
 import org.apache.ignite.internal.processors.cache.persistence.wal.WalEnableDisableWithNodeShutdownTest;
 import org.apache.ignite.internal.processors.cache.persistence.wal.WalEnableDisableWithRestartsTest;
+import org.apache.ignite.internal.processors.cache.persistence.wal.WalPageRecordCompactionTest;
 import org.apache.ignite.internal.processors.cache.persistence.wal.aware.SegmentAwareTest;
 import org.apache.ignite.testframework.GridTestUtils;
 import org.apache.ignite.testframework.junits.DynamicSuite;
@@ -123,6 +124,7 @@ public class IgnitePdsTestSuite5 {
         GridTestUtils.addTestIfNeeded(suite, IgniteCacheDatabaseSharedManagerSelfTest.class, ignoredTests);
 
         GridTestUtils.addTestIfNeeded(suite, WalCompactionNotificationsTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, WalPageRecordCompactionTest.class, ignoredTests);
 
         return suite;
     }
